+++ common/src/main/java/de/cech12/solarcooker/client/SolarCookerRenderState.java
@@ -2,21 +2,14 @@ package de.cech12.solarcooker.client;
 
 import net.minecraft.client.renderer.blockentity.state.BlockEntityRenderState;
 import net.minecraft.client.renderer.item.ItemStackRenderState;
+import net.minecraft.core.Direction;
 
 public class SolarCookerRenderState extends BlockEntityRenderState {
 
-    public float open;
-    public float angle;
-    public boolean hasLeftReflector;
-    public boolean hasRightReflector;
-    public ItemStackRenderState stack;
-
-    public SolarCookerRenderState() {
-        this.open = 0;
-        this.angle = 0;
-        this.hasLeftReflector = false;
-        this.hasRightReflector = false;
-        this.stack = new ItemStackRenderState();
-    }
+    public float open = 0;
+    public Direction facing = Direction.SOUTH;
+    public boolean hasLeftReflector = false;
+    public boolean hasRightReflector = false;
+    public ItemStackRenderState stack = new ItemStackRenderState();
 
 }