+++ common/src/main/java/de/cech12/solarcooker/client/SolarCookerScreen.java
@@ -2,7 +2,7 @@ package de.cech12.solarcooker.client;
 
 import de.cech12.solarcooker.Constants;
 import de.cech12.solarcooker.inventory.SolarCookerContainer;
-import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.renderer.RenderPipelines;
 import net.minecraft.network.chat.Component;
@@ -27,30 +27,28 @@ public class SolarCookerScreen extends AbstractContainerScreen<SolarCookerContai
     }
 
     @Override
-    public void render(@NotNull GuiGraphics guiGraphics, int mouseX, int mouseY, float partialTicks) {
-        this.renderBackground(guiGraphics, mouseX, mouseY, partialTicks);
-        super.render(guiGraphics, mouseX, mouseY, partialTicks);
-        this.renderTooltip(guiGraphics, mouseX, mouseY);
+    public void extractRenderState(@NotNull GuiGraphicsExtractor guiGraphics, int mouseX, int mouseY, float partialTicks) {
+        this.extractBackground(guiGraphics, mouseX, mouseY, partialTicks);
+        super.extractRenderState(guiGraphics, mouseX, mouseY, partialTicks);
+        this.extractTooltip(guiGraphics, mouseX, mouseY);
     }
 
     @Override
-    protected void renderBg(@NotNull GuiGraphics guiGraphics, float partialTicks, int x, int y) {
-        if (this.minecraft != null) {
-            //draw gui
-            int left = this.leftPos;
-            int top = this.topPos;
-            guiGraphics.blit(RenderPipelines.GUI_TEXTURED, guiTexture, left, top, 0.0F, 0.0F, this.imageWidth, this.imageHeight, 256, 256);
-            //draw flame
-            if (this.menu.isBurning()) {
-                guiGraphics.blitSprite(RenderPipelines.GUI_TEXTURED, litSprite, 14, 14, 0, 0, left + 56, top + 36, 14, 14);
-            }
-            //draw progress
-            int progress = this.menu.getCookProgressionScaled();
-            guiGraphics.blitSprite(RenderPipelines.GUI_TEXTURED, burnProgressSprite, 24, 16, 0, 0, left + 79, top + 34, progress, 16);
-            //draw sun
-            if (this.menu.isSunlit()) {
-                guiGraphics.blitSprite(RenderPipelines.GUI_TEXTURED, sunlightSprite, 18, 18, 0, 0, left + 55, top + 52, 18, 18);
-            }
+    public void extractBackground(@NotNull GuiGraphicsExtractor guiGraphics, int mouseX, int mouseY, float partialTicks) {
+        //draw gui
+        int left = this.leftPos;
+        int top = this.topPos;
+        guiGraphics.blit(RenderPipelines.GUI_TEXTURED, guiTexture, left, top, 0.0F, 0.0F, this.imageWidth, this.imageHeight, 256, 256);
+        //draw flame
+        if (this.menu.isBurning()) {
+            guiGraphics.blitSprite(RenderPipelines.GUI_TEXTURED, litSprite, 14, 14, 0, 0, left + 56, top + 36, 14, 14);
+        }
+        //draw progress
+        int progress = this.menu.getCookProgressionScaled();
+        guiGraphics.blitSprite(RenderPipelines.GUI_TEXTURED, burnProgressSprite, 24, 16, 0, 0, left + 79, top + 34, progress, 16);
+        //draw sun
+        if (this.menu.isSunlit()) {
+            guiGraphics.blitSprite(RenderPipelines.GUI_TEXTURED, sunlightSprite, 18, 18, 0, 0, left + 55, top + 52, 18, 18);
         }
     }
 }