+++ neoforge/src/main/java/de/cech12/solarcooker/rei/SolarCookerReiCommonPlugin.java
@@ -1,5 +1,5 @@
 package de.cech12.solarcooker.rei;
-
+/*
 import de.cech12.solarcooker.Constants;
 import de.cech12.solarcooker.crafting.SolarCookingRecipe;
 import de.cech12.solarcooker.platform.Services;
@@ -31,3 +31,4 @@ public class SolarCookerReiCommonPlugin implements REICommonPlugin {
     }
 
 }
+ */